requirements.txt
amqp==5.1.1; python_version >= '3.6'
asgiref==3.5.2; python_version >= '3.7'
astroid==2.12.10; python_full_version >= '3.7.2'
async-timeout==4.0.2; python_version >= '3.6'
attrs==22.1.0
autobahn==22.7.1; python_version >= '3.7'
automat==20.2.0
autopep8==1.7.0
babi==1.5.3
babi-grammars==0.0.45; python_version >= '3.7'
beautifulsoup4==4.11.1; python_version >= '3.6'
billiard==3.6.4.0
black==22.8.0
brotli==1.0.9
celery==5.2.7
cffi==1.15.1
cfgv==3.3.1; python_full_version >= '3.6.1'
channels==3.0.5
click==8.1.3; python_version >= '3.7'
click-didyoumean==0.3.0; python_version < '4.0' and python_full_version >= '3.6.2'
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==38.0.1; python_version >= '3.6'
daphne==3.0.2
deprecated==1.2.13; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
dill==0.3.5.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
distlib==0.3.6
django==4.1.1
django-ajax-datatable==4.4.4
django-appconf==1.0.5; python_version >= '3.6'
django-brotli==0.2.0
django-compressor==4.1
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-extensions==3.2.1
django-filter==22.1
django-htmlmin==0.11.0
django-localflavor==3.1
django-redis==5.2.0
django-redis-cache==1.8.0
django-redis-sessions==0.6.2
django-redisboard==8.3.0
django-stubs-ext==0.5.0; python_version >= '3.6'
django-stubs[compatible-mypy]==1.12.0
django-tz-detect==0.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
filelock==3.8.0; python_version >= '3.7'
flake8==5.0.4
html5lib==1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
hyperlink==21.0.0
identify==2.5.5; python_version >= '3.7'
idna==3.4
incremental==21.3.0
isort==5.10.1; python_version < '4.0' and python_full_version >= '3.6.1'
kombu==5.2.4; python_version >= '3.7'
lazy-object-proxy==1.7.1; python_version >= '3.6'
markdown==3.4.1
mccabe==0.7.0; python_version >= '3.6'
mypy==0.961
mypy-extensions==0.4.3
nodeenv==1.7.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
obd==0.7.1
onigurumacffi==1.2.0; python_version >= '3.7'
packaging==21.3; python_version >= '3.6'
pathspec==0.10.1; python_version >= '3.7'
pillow==9.2.0
pint==0.7.2
platformdirs==2.5.2; python_version >= '3.7'
pre-commit==2.20.0
prompt-toolkit==3.0.31; python_full_version >= '3.6.2'
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1; python_version >= '3.6'
pycparser==2.21
pyflakes==2.5.0; python_version >= '3.6'
pyjwt==2.5.0; python_version >= '3.7'
pylint==2.15.3
pylint-django==2.5.3
pylint-plugin-utils==0.7; python_full_version >= '3.6.2'
pyopenssl==22.0.0
pyparsing==3.0.9; python_full_version >= '3.6.8'
pyserial==3.5
python-stdnum==1.17
pytz==2022.2.1
pyyaml==6.0; python_version >= '3.6'
rcssmin==1.1.0
redis==4.3.4
rjsmin==1.2.0
service-identity==21.1.0
setuptools==65.3.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sorl-thumbnail==12.9.0
soupsieve==2.3.2.post1; python_version >= '3.6'
sqlparse==0.4.2; python_version >= '3.5'
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
tomli==2.0.1; python_full_version < '3.11.0a7'
tomlkit==0.11.4; python_version >= '3.6' and python_version < '4.0'
twisted[tls]==22.8.0; python_full_version >= '3.7.1'
txaio==22.2.1; python_version >= '3.6'
types-pytz==2022.2.1.0
types-pyyaml==6.0.11
typing-extensions==4.3.0; python_version >= '3.7'
vine==5.0.0; python_version >= '3.6'
virtualenv==20.16.5; python_version >= '3.6'
wcwidth==0.2.5
webencodings==0.5.1
whitenoise[brotli]==6.2.0
wrapt==1.14.1; python_version < '3.11'
zope.interface==5.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'

django-braces~=1.15.0
googlemaps~=4.6.0
httpx~=0.23.0
asgiref~=3.5.2